# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateAppleInput:
    properties:
      name:
        type: string
    type: object
  dto.CreatePineAppleInput:
    properties:
      name:
        type: string
    type: object
  dto.UpdateAppleInput:
    properties:
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Microservice for order management with JWT auth, Postgres, Redis, RabbitMQ
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Order Service API
  version: "1.0"
paths:
  /apples/create_apple:
    post:
      consumes:
      - application/json
      description: Create Apple
      parameters:
      - description: Apple creation payload
        in: body
        name: peoduct
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAppleInput'
      produces:
      - application/json
      responses:
        "201":
          description: 'status: Apple created'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: CreateApple
      tags:
      - apple
  /apples/delete_apple/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Apple
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: Apple delete'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: DeleteApple
      tags:
      - apple
  /apples/get_apple/{id}:
    get:
      consumes:
      - application/json
      description: Get Apple
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: Apple'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: GetApple
      tags:
      - apple
  /apples/update_apple/{id}:
    put:
      consumes:
      - application/json
      description: Update Apple
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Apple update payload
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAppleInput'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: Apple'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: UpdateApple
      tags:
      - apple
  /pineapple/create_pineapple:
    post:
      consumes:
      - application/json
      description: Create Pineapple
      parameters:
      - description: Pineapple creation payload
        in: body
        name: peoduct
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePineAppleInput'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: Pineapple'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: CreatePineapple
      tags:
      - pineapple
  /pineapple/get_pineapple/{id}:
    get:
      consumes:
      - application/json
      description: Get Pineapple
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: Pineapple'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: GetPineapple
      tags:
      - pineapple
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"