# internal/product/controller/http_router/v1/create_apple.go
package v1

import (
	"errors"
	"net/http"

	"github.com/G0tem/go-service-layout/internal/product/dto"
	"github.com/G0tem/go-service-layout/internal/product/entity"
	"github.com/G0tem/go-service-layout/pkg/logger"
	"github.com/G0tem/go-service-layout/pkg/otel/tracer"
	"github.com/gin-gonic/gin"
	"go.opentelemetry.io/otel/attribute"
)

// @Summary CreateApple
// @Description Create Apple
// @Tags apple
// @Accept json
// @Produce json
// @Param peoduct body dto.CreateAppleInput true "Apple creation payload"
// @Success 201 {object} map[string]string "status: Apple created"
// @Failure 400 {object} map[string]string "invalid payload"
// @Failure 401 {object} map[string]string "unauthorized"
// @Failure 500 {object} map[string]string "internal error"
// @Router /apples/create_apple [post]
func (h *Handlers) CreateApple(c *gin.Context) {
	ctx, span := tracer.Start(c.Request.Context(), "http/v1 CreateApple")
	defer span.End()

	log := logger.Ctx(c)

	input := dto.CreateAppleInput{}

	err := c.ShouldBindJSON(&input)
	if err != nil {
		log.Error().Err(err).Msg("c.ShouldBindJSON")
		c.JSON(http.StatusBadRequest, gin.H{"error": "json error"})
		return
	}

	err = input.Validate()
	if err != nil {
		log.Error().Err(err).Msg("input.Validate")
		c.JSON(http.StatusBadRequest, gin.H{"error": "validate error"})
		return
	}

	output, err := h.uc.CreateApple(ctx, input)
	if err != nil {
		switch {
		case errors.Is(err, entity.ErrNotFound):
			log.Error().Err(err).Msg("uc.CreateApple: not found")
			c.JSON(http.StatusNotFound, gin.H{"error": "not found"})
			return

		case errors.Is(err, entity.ErrUUIDInvalid), errors.Is(err, entity.ErrStatusInvalid):
			log.Error().Err(err).Msg("uc.CreateApple: validate error")
			c.JSON(http.StatusBadRequest, gin.H{"error": "validate error"})
			return

		default:
			log.Error().Err(err).Msg("uc.CreateApple: internal error")
			c.JSON(http.StatusInternalServerError, gin.H{"error": "internal error"})
			span.SetAttributes(attribute.KeyValue{Key: "error", Value: attribute.StringValue(err.Error())})
			tracer.SetStatus(span, err)
			return
		}
	}

	c.JSON(http.StatusCreated, output)
}